requirements.txt
pytorch
lightning
numpy
pandas
scikit-learn
scipy

optuna
optuna-dashboard
tensorboard
rich